Rename and move the updater
